refactor: Price 값객체 생성
